// swiyu-trust-management-ui/src/app/api/generated/model/task-action.ts
/**
 * SWIYU Trust Management SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const TaskAction = {
  Reject: 'REJECT',
  Approve: 'APPROVE',
  RequestMoreInformation: 'REQUEST_MORE_INFORMATION',
  AddInternalNote: 'ADD_INTERNAL_NOTE',
  AssignSelf: 'ASSIGN_SELF'
} as const;
export type TaskAction = (typeof TaskAction)[keyof typeof TaskAction];